extracted fees related code to eth_fees_oracle
