+++ core/node/eth_sender/src/lib.rs
@@ -9,6 +9,8 @@ mod utils;
 mod zksync_functions;
 
 mod abstract_l1_interface;
+
+mod eth_fees_oracle;
 #[cfg(test)]
 mod tests;
